refactor: 랜딩 페이지 LinkButton 폴더 구조 수정
