+++ frontend/src/pages/landing/HeroSection/HeroSection.styled.ts
