+++ frontend/src/pages/landing/LandingBannerSection/LandingBannerSection.styled.ts
