+++ frontend/src/pages/landing/components/LinkButton/LinkButton.tsx
@@ -1,0 +1,12 @@
+import type { PropsWithChildren } from "react";
+import * as S from "./LinkButton.styled";
+
+interface LinkButton {
+  href: string;
+}
+
+function LinkButton({ children, href }: PropsWithChildren<LinkButton>) {
+  return <S.LinkButton to={href}>{children}</S.LinkButton>;
+}
+
+export default LinkButton;
